Fix false error reporting on todo update response.
jQuery will barf on an json ajax response if there is no actual
response body because it tries to parse the body into json. The
solution is to either set the type of the request to html so jQuery
doesn't try to parse it as json, or return some actual json data in the
reqquest.